Fix for the "MissingPaymentKey" error for minting
Thats a fast fix for the "Missing Payment Key" error when generating a witness with the mintSigningKey for token minting. The filter for mint and multisig was missing and the condition to throw the "Missing Payment" error did not include the presence of a mint key. Please recheck the code, i did a fast fix to test the token minting with a hw-key based policy, which worked afterall.